requirements.txt
cython
numpy
click
pyqt5
opencv-python-headless
pyfirmata
pyqtgraph
pyyaml
